Added Delete Button
Fixed a bug where rename button wasnt taking pre existing names into account.
It now checks if name is available before renaming.
Also seperated functions from routes

+++ main_page.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, Response,request, redirect,url_for
+from flask import Flask, render_template, Response,request, redirect,url_for, flash, send_file
 from flask_sqlalchemy import SQLAlchemy
 import datetime 
 import cv2
@@ -43,6 +43,40 @@ class Record(db.Model):
 		x = datetime.datetime.now()
 		self.video_name = name+str(count)+'_'+str(x.date())+'_'+x.strftime("%H")+'.'+x.strftime("%M")+'.'+x.strftime("%S")
 
+
+# ===================== Functions ====================
+
+
+def gen_frames(): 
+	while True:
+		success, frame = camera.read()  # read the camera frame
+		if not success:
+			break
+		else:
+			ret, buffer = cv2.imencode('.jpg', frame)
+			frame = buffer.tobytes()
+			yield (b'--frame\r\n'
+				b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')  # concat frame one by one and show result
+
+# Generating the thumbnail is slow needs improving
+def gen_thumbnail(video_name): 
+	video = cv2.VideoCapture(f'static/{video_name}.mp4')
+	if video.isOpened():
+		video.set(2,0.5); # Set Frame to middle
+		success, frame = video.retrieve()  # read the camera frame
+		video.release()
+		frame = cv2.resize(frame, (52,30))
+		ret, buffer = cv2.imencode('.jpg', frame)
+		frame = buffer.tobytes()
+		yield (b'--frame\r\n'
+			b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')  # concat frame one by one and show result
+	else:
+		video.release()
+		yield(b'--frame\r\n')
+
+
+
+
 # ===================== Routes ====================
 
 
@@ -53,22 +87,25 @@ def index():
 	
 	#if "open" in request.form:
 		
-	
 	if initi:
-		for i in range(10):
+		for i in range(5):
 			count = Record.query.count() 
 			new_record = Record("Video",count)
 			db.session.add(new_record)
 		db.session.commit()
 		initi = False
 
-	
-	
-
-
 	records = Record.query.order_by(Record.id.desc()).all()
 	#print("Test: ",current_video, file=sys.stderr)
-	return render_template('index.html',video_name = current_video, current_video_id=current_video_id, video_format = video_format,records = records )
+	return render_template('index.html',video_name = current_video, 
+		current_video_id=current_video_id, video_format = video_format,records = records)
+
+
+@app.route('/download/<video_name>,<video_format>')
+def download_video (video_name,video_format):
+    #For windows you need to use drive name [ex: F:/Example.pdf]
+    path = f"static/{video_name}.{video_format}"
+    return send_file(path, as_attachment=True)
 
 
 @app.route('/delete_video/<video_id>')
@@ -87,10 +124,10 @@ def delete_video(video_id):
 @app.route('/rename_video/<video_id>',methods=['POST'])
 def rename_video(video_id):
 	global current_video
-
 	if request.method == "POST":
-		#print(form.name, file=sys.stderr)
 		new_name = request.form["new_video_name"];
+		is_name_available = db.session.query(Record.id).filter_by(video_name=new_name).first() is not None
+		if is_name_available: flash(f" ERROR: '{new_name}' already exists!"); return redirect('/')
 		# Rename Video from Server
 		if os.path.isfile(f'static/{current_video}.{video_format}'):
 			os.rename(f'static/{current_video}.{video_format}',f'static/{new_name}.{video_format}')
@@ -109,35 +146,6 @@ def play_video(video_name,video_id):
 	current_video = video_name
 	current_video_id = video_id
 	return redirect("/")
-
-
-
-def gen_frames(): 
-	while True:
-		success, frame = camera.read()  # read the camera frame
-		if not success:
-			break
-		else:
-			ret, buffer = cv2.imencode('.jpg', frame)
-			frame = buffer.tobytes()
-			yield (b'--frame\r\n'
-				b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')  # concat frame one by one and show result
-
-
-def gen_thumbnail(video_name): 
-	video = cv2.VideoCapture(f'static/{video_name}.mp4')
-	if video.isOpened():
-		video.set(2,0.5); # Set Frame to middle
-		success, frame = video.retrieve()  # read the camera frame
-		video.release()
-		frame = cv2.resize(frame, (52,30))
-		ret, buffer = cv2.imencode('.jpg', frame)
-		frame = buffer.tobytes()
-		yield (b'--frame\r\n'
-			b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')  # concat frame one by one and show result
-	else:
-		video.release()
-		yield(b'--frame\r\n')
 
 @app.route('/get_thumbnail/<video_name>')
 def get_thumbnail(video_name):
